chore(PPDSC-2329): chore-ppdse-2329-moved utils to utils folder
