+++ lib/codemod/src/transforms/update-linkstandalone.js
@@ -1,4 +1,4 @@
-const {replaceTags, replaceImports} = require('../utils');
+const {replaceTags, replaceImports} = require('../utils/component-utils');
 
 //convert Link to LinkStandalone
