+++ lib/codemod/src/utils/component-utils.js
@@ -1,4 +1,4 @@
-// Functions for changing Link to LinkInline /LinkStandalone
+// Functions for changing component tags and imports
 
 function findTags(j, oldTag, newTag) {
   const name = oldTag.find(j.JSXIdentifier, {name: 'Link'});
